\name{MysterySamples-package}
\alias{MysterySamples-package}
\alias{MysterySamples}
\docType{package}
\title{
An introduction to microarray analysis using R and Bioconductor
}
\description{
This package uses basic R/Bioconductor infrastructure and functionality to perform quality control, preprocessing, exploratory data analysis, and differential expression on a set of 10 samples.  It is based on publicly available one-color oligonucleotide expression data from ten samples, two each from five different normal tissues. The task is to follow the vignette and identify the five unknown tissues.
}
\details{
\tabular{ll}{
Package: \tab MysterySamples\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-06-15\cr
License: \tab MIT\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Sean Davis

Maintainer: Sean Davis <sdavis2@mail.nih.gov>
}
\references{
Su AI, Wiltshire T, Batalov S, Lapp H, Ching KA, Block D, Zhang J, Soden R,
Hayakawa M, Kreiman G, Cooke MP, Walker JR, Hogenesch JB. A gene atlas of the
mouse and human protein-encoding transcriptomes. Proc Natl Acad Sci U S A. 2004
Apr 20;101(16):6062-7. Epub 2004 Apr 9. PubMed PMID: 15075390; PubMed Central
PMCID: PMC395923.
}
\keyword{ package }
\examples{
data(samples)

}
